requirements.txt
customtkinter
SpeechRecognition
pyaudio
keyboard
Pillow
pyperclip
